2.0 - Skullgirls - New Humble installer + More robust postinst/prerm scripts

+++ play.it-2/games/play-skullgirls.sh
@@ -34,29 +34,37 @@ set -o errexit
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20170523.1
+script_version=20170708.1
 
 # Set game-specific variables
 
 GAME_ID='skullgirls'
 GAME_NAME='Skullgirls'
 
-ARCHIVES_LIST='ARCHIVE_HUMBLE'
+ARCHIVES_LIST='ARCHIVE_HUMBLE ARCHIVE_HUMBLE_OLD'
 
-ARCHIVE_HUMBLE='Skullgirls-1.0.1.sh'
-ARCHIVE_HUMBLE_MD5='bf110f7d29bfd4b9e075584e41fef402'
-ARCHIVE_HUMBLE_TYPE='mojosetup'
+ARCHIVE_HUMBLE='Skullgirls-15719.tar'
+ARCHIVE_HUMBLE_MD5='104a6976aec70d423756e008a5b8554c'
 ARCHIVE_HUMBLE_SIZE='4200000'
-ARCHIVE_HUMBLE_VERSION='1.0.1-humble152310'
+ARCHIVE_HUMBLE_VERSION='15719-humble170628'
 
-ARCHIVE_GAME_DATA_PATH='data/noarch'
-ARCHIVE_GAME_DATA_FILES='./*'
+ARCHIVE_HUMBLE_OLD='Skullgirls-1.0.1.sh'
+ARCHIVE_HUMBLE_OLD_MD5='bf110f7d29bfd4b9e075584e41fef402'
+ARCHIVE_HUMBLE_OLD_TYPE='mojosetup'
+ARCHIVE_HUMBLE_OLD_SIZE='4200000'
+ARCHIVE_HUMBLE_OLD_VERSION='1.0.1-humble152310'
 
-ARCHIVE_GAME_BIN32_PATH='data/i686'
-ARCHIVE_GAME_BIN32_FILES='./*'
+ARCHIVE_GAME_BIN32_PATH_HUMBLE='SkullGirls'
+ARCHIVE_GAME_BIN32_PATH_HUMBLE_OLD='data/i686'
+ARCHIVE_GAME_BIN32_FILES='./SkullGirls.i686-pc-linux-gnu ./lib/i686-pc-linux-gnu'
 
-ARCHIVE_GAME_BIN64_PATH='data/x86_64'
-ARCHIVE_GAME_BIN64_FILES='./*'
+ARCHIVE_GAME_BIN64_PATH_HUMBLE='SkullGirls'
+ARCHIVE_GAME_BIN64_PATH_HUMBLE_OLD='data/x86_64'
+ARCHIVE_GAME_BIN64_FILES='./SkullGirls.x86_64-pc-linux-gnu ./lib/x86_64-pc-linux-gnu'
+
+ARCHIVE_GAME_DATA_PATH_HUMBLE='SkullGirls'
+ARCHIVE_GAME_DATA_PATH_HUMBLE_OLD='data/noarch'
+ARCHIVE_GAME_DATA_FILES='./*.txt ./data01 ./Icon.png ./Salmon'
 
 APP_MAIN_TYPE='native'
 APP_MAIN_EXE_BIN32='SkullGirls.i686-pc-linux-gnu'
@@ -71,11 +79,11 @@ PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN32_ARCH='32'
 PKG_BIN32_DEPS_DEB="$PKG_DATA_ID, libc6, libstdc++6, libsdl2-mixer-2.0-0, libsdl2-2.0-0"
-PKG_BIN32_DEPS_ARCH="$PKG_DATA_ID lib32-sdl2 lib32-sdl2_mixer"
+PKG_BIN32_DEPS_ARCH="$PKG_DATA_ID lib32-glibc lib32-gcc-libs lib32-sdl2 lib32-sdl2_mixer"
 
 PKG_BIN64_ARCH='64'
 PKG_BIN64_DEPS_DEB="$PKG_BIN32_DEPS_DEB"
-PKG_BIN64_DEPS_ARCH="$PKG_DATA_ID sdl2 sdl2_mixer"
+PKG_BIN64_DEPS_ARCH="$PKG_DATA_ID glibc gcc-libs sdl2 sdl2_mixer"
 
 # Load common functions
 
@@ -122,13 +130,17 @@ res="$APP_MAIN_ICON_RES"
 PATH_ICON="$PATH_ICON_BASE/${res}x${res}/apps"
 
 cat > "$postinst" << EOF
-mkdir --parents "$PATH_ICON"
-ln --symbolic "$PATH_GAME"/$APP_MAIN_ICON "$PATH_ICON/$GAME_ID.png"
+if ! [ -e "$PATH_ICON/$GAME_ID.png" ]; then
+	mkdir --parents "$PATH_ICON"
+	ln --symbolic "$PATH_GAME"/$APP_MAIN_ICON "$PATH_ICON/$GAME_ID.png"
+fi
 EOF
 
 cat > "$prerm" << EOF
-rm "$PATH_ICON/$GAME_ID.png"
-rmdir --parents --ignore-fail-on-non-empty "$PATH_ICON"
+if [ -e "$PATH_ICON/$GAME_ID.png" ]; then
+	rm "$PATH_ICON/$GAME_ID.png"
+	rmdir --parents --ignore-fail-on-non-empty "$PATH_ICON"
+fi
 EOF
 
 write_metadata 'PKG_DATA'
